requirements.txt
streamlit
gspread
google-auth
pandas
transformers
torch
